Added db_access.py
Created cloudant client that uploads all of the json produced by
get_courses.py stored in a local file and uploads it to 'testing_db' on
Cloudant. We should probably rename testing_db to dev_db as a staged db
service that we then replicate to the purdue_courses database.